Add paints page
Added the brand name under the colors in the paints tab

added a group radio to select the brand you prefer

added in a basic color chart to select a color

added in the add paint button.

I still need to get the info from the paints-form page to create a new object on the paints tab.